+++ RELEASE-NOTES.md
@@ -41,6 +41,7 @@
   commandline flag `-enable_revocable_ram`.
 - Introduce UpdateMetadata fields in JobUpdateRequest to allow clients to store metadata on update.
 - Changed cronSchedule field inside of JobConfiguration schema to be optional for compatibility with Go.
+- Update default value of command line option `-framework_name` to 'Aurora'.
 
 ### Deprecations and removals: